+++ Algorithms/Searching & Sorting/QuickSort/quicksort.java
@@ -1,48 +1,45 @@
-#include<stdio.h>
+class QuickSort { 
+    static int partition(int elem[], int low, int high) { 
+        int pivot = elem[high];  
+        int i = (low-1); // index of smaller element 
+        for (int j=low; j<high; j++)
+            if (elem[j] < pivot) { // check scale for elements less than pivot
+                i++; 
+                int temp = elem[i]; 
+                elem[i] = elem[j]; 
+                elem[j] = temp;  // swap
+            } 
 
-void quicksort(int num[25],int f,int l){
-   int i, j, pivot, temp;
+        int temp = elem[i+1]; 
+        elem[i+1] = elem[high]; 
+        elem[high] = temp; // swap
+        return i+1; 
+    } 
+  
+    static void quickSort(int elem[], int low, int high) { 
+        if (low < high) { 
+            int pi = partition(elem, low, high); 
+            quickSort(elem, low, pi-1); 
+            quickSort(elem, pi+1, high); 
+        } 
+    } 
+  
+    // result_print
+    static void print_result(int elem[]) {
+        int n = elem.length; 
+        for (int i=0; i<n; ++i) 
+            System.out.print(elem[i]+" "); 
+    } 
+  
+    // Driver function
+    public static void main(String args[]) { 
+        int elem[] = {9, 7, 2, 3, 7, 11, 01, 90, 12, 82}; 
+        int n = elem.length; 
+  
+        QuickSort qsj = new QuickSort(); 
+        qsj.quickSort(elem, 0, n-1);
 
-   if(f<l){
-      pivot=f;
-      i=f;
-      j=l;
-
-      while(i<j){
-         while(num[i]<=num[pivot]&&i<l)
-            i++;
-         while(num[j]>num[pivot])
-            j--;
-         if(i<j){
-            temp=num[i];
-            num[i]=num[j];
-            num[j]=temp;
-         }
-      }
-
-      temp=num[pivot];
-      num[pivot]=num[j];
-      num[j]=temp;
-      quicksort(num,f,j-1);
-      quicksort(num,j+1,l);
-
-   }
-}
-
-void main(){
-   int i, count, num[25];
-
-   printf("How many elements are u going to enter?: ");
-   scanf("%d",&count);
-
-   printf("Enter %d elements: ", count);
-   for(i=0;i<count;i++)
-      scanf("%d",&num[i]);
-
-   quicksort(num,0,count-1);     //function call
-
-   printf("Order of Sorted elements: ");
-   for(i=0;i<count;i++)
-      printf(" %d",num[i]);
-
+        System.out.println("The sorted array is:"); 
+        print_result(elem); 
+    } 
 }
